\name{alphashape3d-package}
\alias{alphashape3d-package}
\docType{package}
\title{
Computation of the 3D \eqn{\alpha}-shape}
\description{
The package alphashape3d presents the implementation in R of the \eqn{\alpha}-shape of a finite set of points in the three-dimensional space. This geometric structure generalizes the convex hull and allows to recover the shape of non-convex and even non-connected sets in 3D, given a random sample of points taken into it. Besides the computation of the \eqn{\alpha}-shape, the package alphashape3d provides users with functions to compute the volume of the \eqn{\alpha}-shape, identify the connected components and facilitate the three-dimensional graphical visualization of the estimated set. 
}
\details{
\tabular{ll}{
Package: \tab alphashape3d\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-05-19\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Thomas Lafarge, Beatriz Pateiro-Lopez.

Maintainers: Beatriz Pateiro-Lopez <beatriz.pateiro@usc.es>
}
\references{
Edelsbrunner, H., Mucke, E. P. (1994). Three-Dimensional Alpha Shapes. \emph{ACM Transactions on Graphics}, 13(1), pp.43-72.
}
\keyword{package}
